Simplify project secret behaviour
This removes the complicated old behaviour where the secret could be
*either* the event secret, or a secret just for the project.

Now, the secret is always on the project - we don't allow a secret to
be set for the event anymore.

This also includes simpler methods to check the secret when a project
is updated, which doesn't require using conditional validation methods.